// types/vehicle.ts
export interface Vehicle {
  id: string;
  year: number;
  brand: string;
  model: string;
  type: string;
  img: string;
  description: string;
  fuelConsumption: string;
  engineSize: string;

  accessories: string[];
  functionalities: string[];

  rentalPrice: string;
  rentalCompany: string;
  address: string;

  rentalConditions: string[];
  mileage: number;
}

export interface VehiclesResponse {
  cars: Vehicle[];
  totalCars: number;
  page: number;
  totalPages: number;
}